[REF] rename module and improve tests;
