--- pos_slate/i18n/fr.po
@@ -1,55 +1,0 @@
-# Translation of OpenERP Server.
-# This file contains the translation of the following modules:
-#	* pos_slate
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: OpenERP Server 7.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2015-03-10 14:52+0000\n"
-"PO-Revision-Date: 2015-03-10 14:52+0000\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: pos_slate
-#: field:pos.config,allow_slate:0
-msgid "Allow Slate"
-msgstr "Autoriser les ardoises"
-
-#. module: pos_slate
-#: code:addons/pos_slate/model/pos_session.py:43
-#, python-format
-msgid "Error!"
-msgstr "Erreur !"
-
-#. module: pos_slate
-#: help:pos.config,allow_slate:0
-msgid "If you check this field, users will have the possibility to let some PoS orders in the slate, and allow the customer to paid later.\n"
-"Order in the slate will not generate entries during the close of the session."
-msgstr "Si vous cochez cette case, les utilisateurs auront la possibilité de laisser certaines vente en ardoise, et autorisera le client à payer plus tard.\n les ventes en ardoise ne génère pas d'écriture lors de la fermeture de la session."
-
-#. module: pos_slate
-#: field:pos.order,is_partial_paid:0
-msgid "Is Partially Paid"
-msgstr "Est partiellement payé"
-
-#. module: pos_slate
-#: code:_description:0
-#: model:ir.model,name:pos_slate.model_pos_order
-#, python-format
-msgid "Point of Sale"
-msgstr "Point de Vente"
-
-#. module: pos_slate
-#: code:addons/pos_slate/model/pos_session.py:46
-#, python-format
-msgid "You cannot confirm this session, because '%s' is in a 'draft' state with payments.\n"
-"\n"
-"Please finish to pay this Order."
-msgstr "Vous ne pouvez pas fermer cette session car '%s' est en brouillon avec des paiments.\n"
-"\n"
-"Merci de compléter les paiments."
