--- pos_slate/tests/__init__.py
@@ -1,2 +1,0 @@
-# -*- encoding: utf-8 -*-
-from . import tests
