+++ pos_store_draft_order/demo/pos_config.yml
@@ -1,7 +1,7 @@
 # -*- encoding: utf-8 -*-
 ##############################################################################
 #
-#    Point Of Sale - Slate module for Odoo
+#    Point Of Sale - Store Draft Orders Module for Odoo
 #    Copyright (C) 2015-Today GRAP (http://www.grap.coop)
 #    @author Sylvain LE GAL (https://twitter.com/legalsylvain)
 #
@@ -21,4 +21,4 @@
 ##############################################################################
 
 - !record {model: pos.config, id: point_of_sale.pos_config_main}:
-    allow_slate: True
+    allow_store_draft_order: True
