+++ pos_store_draft_order/tests/__init__.py
@@ -1,0 +1,2 @@
+# -*- encoding: utf-8 -*-
+from . import test_pos_store_draft_order
