+++ pos_store_draft_order/tests/test_pos_store_draft_order.py
@@ -1,7 +1,7 @@
 # -*- encoding: utf-8 -*-
 ##############################################################################
 #
-#    Point Of Sale - Slate module for Odoo
+#    Point Of Sale - Store Draft Orders Module for Odoo
 #    Copyright (C) 2015-Today GRAP (http://www.grap.coop)
 #    @author Sylvain LE GAL (https://twitter.com/legalsylvain)
 #
@@ -24,14 +24,13 @@ from openerp.exceptions import Warning
 from openerp.tests.common import TransactionCase
 
 
-class TestPointOfSaleSlate(TransactionCase):
-    """Tests for 'Point Of Sale - Slate' Module"""
+class TestPosStoreDraftOrder(TransactionCase):
 
     def setUp(self):
-        super(TestPointOfSaleSlate, self).setUp()
+        super(TestPosStoreDraftOrder, self).setUp()
 
-        self.session_obj = self.registry('pos.session')
-        self.order_obj = self.registry('pos.order')
+        self.session_obj = self.env['pos.session']
+        self.order_obj = self.env['pos.order']
 
         self.pos_config_id = self.ref('point_of_sale.pos_config_main')
         self.product_id = self.ref('product.product_product_48')
@@ -39,8 +38,8 @@ class TestPointOfSaleSlate(TransactionCase):
 
     # Test Section
     def test_01_allow_draft_order_unpaid(self):
-        """Test the possibility to let a PoS Order in a slate if it is not
-        paid at all."""
+        """Test the possibility to let a PoS Order in a draft state if it is
+         not paid at all."""
         # Open a new session
         session_1 = self.session_obj.create({'config_id': self.pos_config_id})
 
@@ -54,14 +53,9 @@ class TestPointOfSaleSlate(TransactionCase):
                 'qty': 1}]]
         })
 
-        # Close Session # TODO TEST
+        # Close Session
         session_1.signal_workflow('close')
-#        self.wf_service.trg_validate(
-#            uid, 'pos.session', session_1.id, 'close', cr)
-
-        # TODO TEST WITHOUT
-#        ps1 = self.session_obj.browse(cr, uid, ps1_id)
-
+        
         self.assertEquals(
             session_1.state, 'closed',
             "Unpaid Draft Orders must not block the closing process of the"
@@ -70,9 +64,6 @@ class TestPointOfSaleSlate(TransactionCase):
         # Open a second session
         session_2 = self.session_obj.create({'config_id': self.pos_config_id})
 
-        # TODO TEST WITHOUT
-#        order = self.order_obj.browse(order.id)
-
         self.assertEquals(
             order.session_id.id, session_2.id,
             "Draft Orders of a previous session must be associated to the"
@@ -80,7 +71,7 @@ class TestPointOfSaleSlate(TransactionCase):
 
     # Test Section
     def test_02_block_draft_order_partial_paid(self):
-        """Test the unpossibility to let a PoS Order in a slate if it is
+        """Test the unpossibility to let a PoS Order in a draft state if it is
         in a partial paid state."""
         # Open a new session
         session_1 = self.session_obj.create({'config_id': self.pos_config_id})
@@ -96,13 +87,9 @@ class TestPointOfSaleSlate(TransactionCase):
         })
 
         # Make partial payment
-        order.add_payment({
-            'amount': 1,
-            'journal': self.cash_journal_id,
-        })
+        self.order_obj.add_payment(
+            order.id, {'amount': 1, 'journal': self.cash_journal_id})
 
         # Try Close Session (Must fail)
         with self.assertRaises(Warning):
             session_1.signal_workflow('close')
-#            self.wf_service.trg_validate(
-#                uid, 'pos.session', ps_id, 'close', cr)
